+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Refactored variables for easier installation (now requiring no parameters to the install script), better reuse of code, and correctly namespaced items
     * `SHOW_DOTS_MESSAGE` is now `DOTS_SHOW_INIT_MESSAGE`
     * `DOTFILES_URL` is no longer used anywhere and no longer needed (this also means the "Powered by ... Dotfiles" message is gone)
+* Added `dots_update` which will update Dots in your dotfiles project
 * Fixed a bug that would throw an error if no shell config file was found on startup (this was problematic as we started assuming the user already had a shell config file; however, that may not be the case for a brand new machine getting setup with Dots)
 * Various other improvements made to the underlyinng code including refactor, comments, updated documentation, etc